+++ WooCommerce/src/main/kotlin/com/woocommerce/android/AppPrefs.kt
@@ -31,6 +31,7 @@ import com.woocommerce.android.tools.SelectedSite
 import com.woocommerce.android.ui.payments.cardreader.onboarding.PersistentOnboardingData
 import com.woocommerce.android.ui.payments.cardreader.onboarding.PluginType
 import com.woocommerce.android.ui.products.ProductType
+import com.woocommerce.android.ui.promo_banner.PromoBannerType
 import com.woocommerce.android.util.PreferenceUtils
 import com.woocommerce.android.util.ThemeOption
 import com.woocommerce.android.util.ThemeOption.DEFAULT
@@ -355,6 +356,17 @@ object AppPrefs {
 
     private fun getReceiptKey(localSiteId: Int, remoteSiteId: Long, selfHostedSiteId: Long, orderId: Long) =
         PrefKeyString("$RECEIPT_PREFIX:$localSiteId:$remoteSiteId:$selfHostedSiteId:$orderId")
+
+    private fun getPromoBannerKey(bannerType: PromoBannerType) =
+        PrefKeyString("PROMO_BANNER_SHOWN_${bannerType}")
+
+    fun hasPromoBannerShown(bannerType: PromoBannerType): Boolean {
+        return getBoolean(getPromoBannerKey(bannerType), false)
+    }
+
+    fun setPromoBannerShown(bannerType: PromoBannerType, shown: Boolean) {
+        setBoolean(getPromoBannerKey(bannerType), shown)
+    }
 
     fun setLastConnectedCardReaderId(readerId: String) =
         setString(UndeletablePrefKey.LAST_CONNECTED_CARD_READER_ID, readerId)
